# public/openapi.yaml
openapi: 3.1.0
info:
  title: Conduit API Gateway
  version: 1.0.0
  description: |
    Conduit is a transparent, edge-first API gateway and proxy for Claude AI.

    ## Features
    - API Key Management with rotation and expiration
    - Multi-provider support with automatic failover
    - Rate limiting and quota enforcement
    - IP whitelisting and HMAC request signing
    - Response caching for improved performance
    - Multi-tenancy with organizations and projects
    - Comprehensive analytics and cost projections
    - Audit logging and 2FA authentication

  contact:
    name: Conduit Support
    url: https://conduit.dev
  license:
    name: MIT
servers:
  - url: https://api.conduit.dev
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/admin/keys:
    get:
      tags: [API Keys]
      summary: List API keys
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: organizationId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      tags: [API Keys]
      summary: Create API key
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - providerId
              properties:
                name:
                  type: string
                providerId:
                  type: string
                  format: uuid
                organizationId:
                  type: string
                  format: uuid
                projectId:
                  type: string
                  format: uuid
                requestsPerMinute:
                  type: integer
                requestsPerDay:
                  type: integer
                tokensPerDay:
                  type: integer
                monthlySpendLimitUsd:
                  type: integer
                  description: Spend limit in cents
                expiresAt:
                  type: string
                  format: date-time
                scopes:
                  type: object
                  properties:
                    models:
                      type: array
                      items:
                        type: string
                    endpoints:
                      type: array
                      items:
                        type: string
                ipWhitelist:
                  type: array
                  items:
                    type: string
                  description: IP addresses or CIDR ranges
                ipBlacklist:
                  type: array
                  items:
                    type: string
                hmacEnabled:
                  type: boolean
                emailNotificationsEnabled:
                  type: boolean
                notificationEmail:
                  type: string
                  format: email
                webhookUrl:
                  type: string
                  format: uri
                slackWebhook:
                  type: string
                  format: uri
                discordWebhook:
                  type: string
                  format: uri
                alertThresholds:
                  type: object
                  properties:
                    requestsPerDay:
                      type: array
                      items:
                        type: integer
                    tokensPerDay:
                      type: array
                      items:
                        type: integer
                    monthlySpend:
                      type: array
                      items:
                        type: integer
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                    description: The full API key (only shown once)
                  keyData:
                    $ref: '#/components/schemas/ApiKey'
...
  /api/admin/organizations:
    get:
      tags: [Organizations]
      summary: List organizations
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of organizations
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Organization'
    post:
      tags: [Organizations]
      summary: Create organization
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - slug
              properties:
                name:
                  type: string
                slug:
                  type: string
                  pattern: '^[a-z0-9-]+$'
                plan:
                  type: string
                  enum: [free, pro, enterprise]
                maxApiKeys:
                  type: integer
                maxUsers:
                  type: integer
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema:
                type: object
                properties:
                  organization:
                    $ref: '#/components/schemas/Organization'
  /api/admin/projects:
    get:
      tags: [Projects]
      summary: List projects
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
    post:
      tags: [Projects]
      summary: Create project
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - organizationId
                - name
              properties:
                organizationId:
                  type: string
                  format: uuid
                name:
                  type: string
                description:
                  type: string
                sharedQuotas:
                  type: object
      responses:
        '201':
          description: Project created
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Conduit API key (sk-cond_...)
    HmacAuth:
      type: apiKey
      in: header
      name: x-signature
      description: HMAC signature for request signing
    BearerAuth:
      type: http
      scheme: bearer
      description: Admin session token
  schemas:
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        keyPrefix:
          type: string
          example: "sk-cond_abc"
        name:
          type: string
        providerId:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        requestsPerMinute:
          type: integer
        requestsPerDay:
          type: integer
        tokensPerDay:
          type: integer
        monthlySpendLimitUsd:
          type: integer
        expiresAt:
          type: string
          format: date-time
        lastRotatedAt:
          type: string
          format: date-time
        rotationCount:
          type: integer
        scopes:
          type: object
        ipWhitelist:
          type: array
          items:
            type: string
        ipBlacklist:
          type: array
          items:
            type: string
        hmacEnabled:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
          enum: [free, pro, enterprise]
        maxApiKeys:
          type: integer
        maxUsers:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        sharedQuotas:
          type: object
        createdAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer