# src/app/api/admin/organizations/route.ts
import { NextRequest, NextResponse } from 'next/server';
import { auth } from '@/lib/auth';
import { db } from '@/lib/db';
import { organizations } from '@/lib/db/schema';
import { logAudit } from '@/lib/audit';

export const runtime = 'edge';

// GET /api/admin/organizations - List all organizations
export async function GET() {
  try {
    const session = await auth();
    if (!session?.user?.email) {
      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
    }

    const orgs = await db.select().from(organizations).orderBy(organizations.createdAt);

    return NextResponse.json({ organizations: orgs });
  } catch (error) {
    console.error('Failed to fetch organizations:', error);
    return NextResponse.json(
      { error: 'Failed to fetch organizations' },
      { status: 500 }
    );
  }
}

// POST /api/admin/organizations - Create a new organization
export async function POST(request: NextRequest) {
  try {
    const session = await auth();
    if (!session?.user?.email) {
      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
    }

    const body = await request.json();
    const { name, slug, plan, maxApiKeys, maxUsers } = body;

    if (!name || !slug) {
      return NextResponse.json(
        { error: 'Name and slug are required' },
        { status: 400 }
      );
    }

    const [org] = await db
      .insert(organizations)
      .values({
        name,
        slug,
        plan: plan || 'free',
        maxApiKeys: maxApiKeys || 10,
        maxUsers: maxUsers || 5,
      })
      .returning();

    // Log audit
    await logAudit({
      adminEmail: session.user.email,
      resourceType: 'organization',
      resourceId: org.id,
      action: 'create',
      ipAddress: request.headers.get('x-forwarded-for') || undefined,
    });

    return NextResponse.json({ organization: org }, { status: 201 });
  } catch (error: unknown) {
    const err = error as { code?: string };
    console.error('Failed to create organization:', error);

    if (err.code === '23505') {
      // Unique constraint violation
      return NextResponse.json(
        { error: 'Organization slug already exists' },
        { status: 409 }
      );
    }

    return NextResponse.json(
      { error: 'Failed to create organization' },
      { status: 500 }
    );
  }
}